+++ skbuild/command/bdist_wheel.py
@@ -9,4 +9,8 @@ from ..utils import new_style
 
 class bdist_wheel(set_build_base_mixin, new_style(_bdist_wheel)):
     """Custom implementation of ``bdist_wheel`` setuptools command."""
-    pass
+
+    def write_wheelfile(self, wheelfile_base, _=None):
+        from .. import __version__ as skbuild_version
+        generator = "skbuild %s" % skbuild_version
+        super(bdist_wheel, self).write_wheelfile(wheelfile_base, generator)